# [Internal] Thin Client Integration: Adds thinclient mode and performance tests.