--- a/Microsoft.Azure.Cosmos/src/DocumentClient.cs
+++ b/Microsoft.Azure.Cosmos/src/DocumentClient.cs
@@ -115,6 +115,10 @@ namespace Microsoft.Azure.Cosmos
         private readonly bool IsLocalQuorumConsistency = false;
         private readonly bool isReplicaAddressValidationEnabled;
 
+        // Thin Client
+        private readonly bool isLiteClientEnabled;
+        private readonly string liteClientEndpoint;
+
         //Fault Injection
         private readonly IChaosInterceptorFactory chaosInterceptorFactory;
         private readonly IChaosInterceptor chaosInterceptor;
@@ -162,6 +166,8 @@ namespace Microsoft.Azure.Cosmos
         // creator of TransportClient is responsible for disposing it.
         private IStoreClientFactory storeClientFactory;
         internal CosmosHttpClient httpClient { get; private set; }
+
+        internal CosmosHttpClient liteModeHttpClient { get; private set; }
 
         // Flag that indicates whether store client factory must be disposed whenever client is disposed.
         // Setting this flag to false will result in store client factory not being disposed when client is disposed.
@@ -241,6 +247,12 @@ namespace Microsoft.Azure.Cosmos
             this.Initialize(serviceEndpoint, connectionPolicy, desiredConsistencyLevel);
             this.initTaskCache = new AsyncCacheNonBlocking<string, bool>(cancellationToken: this.cancellationTokenSource.Token);
             this.isReplicaAddressValidationEnabled = ConfigurationManager.IsReplicaAddressValidationEnabled(connectionPolicy);
+            this.isLiteClientEnabled = ConfigurationManager.IsLiteClientEnabled(defaultValue: true);
+
+            if (this.isLiteClientEnabled)
+            {
+                this.liteClientEndpoint = ConfigurationManager.GetLiteClientEndpoint(defaultValue: "https://jsonplaceholder.typicode.com/posts");
+            }
         }
 
         /// <summary>
@@ -493,6 +505,12 @@ namespace Microsoft.Azure.Cosmos
             this.initTaskCache = new AsyncCacheNonBlocking<string, bool>(cancellationToken: this.cancellationTokenSource.Token);
             this.chaosInterceptorFactory = chaosInterceptorFactory;
             this.chaosInterceptor = chaosInterceptorFactory?.CreateInterceptor(this);
+            this.isLiteClientEnabled = ConfigurationManager.IsLiteClientEnabled(defaultValue: true);
+
+            if (this.isLiteClientEnabled)
+            {
+                this.liteClientEndpoint = ConfigurationManager.GetLiteClientEndpoint(defaultValue: "https://jsonplaceholder.typicode.com/posts");
+            }
 
             this.Initialize(
                 serviceEndpoint: serviceEndpoint,
@@ -504,7 +522,8 @@ namespace Microsoft.Azure.Cosmos
                 storeClientFactory: storeClientFactory,
                 cosmosClientId: cosmosClientId,
                 remoteCertificateValidationCallback: remoteCertificateValidationCallback,
-                cosmosClientTelemetryOptions: cosmosClientTelemetryOptions);
+                cosmosClientTelemetryOptions: cosmosClientTelemetryOptions,
+                enableLiteClientMode: this.isLiteClientEnabled);
         }
 
         /// <summary>
@@ -688,7 +707,8 @@ namespace Microsoft.Azure.Cosmos
             TokenCredential tokenCredential = null,
             string cosmosClientId = null,
             RemoteCertificateValidationCallback remoteCertificateValidationCallback = null,
-            CosmosClientTelemetryOptions cosmosClientTelemetryOptions = null)
+            CosmosClientTelemetryOptions cosmosClientTelemetryOptions = null,
+            bool enableLiteClientMode = false)
         {
             if (serviceEndpoint == null)
             {
@@ -947,6 +967,17 @@ namespace Microsoft.Azure.Cosmos
                 this.sendingRequest,
                 this.receivedResponse);
 
+            if (enableLiteClientMode)
+            {
+                this.liteModeHttpClient = CosmosHttpClientCore.CreateWithConnectionPolicy(
+                    this.ApiType,
+                    DocumentClientEventSource.Instance,
+                    this.ConnectionPolicy,
+                    null,
+                    this.sendingRequest,
+                    this.receivedResponse);
+            }
+
             // Loading VM Information (non blocking call and initialization won't fail if this call fails)
             VmMetadataApiHandler.TryInitialize(this.httpClient);
 
@@ -1051,6 +1082,19 @@ namespace Microsoft.Azure.Cosmos
             if (this.ConnectionPolicy.ConnectionMode == ConnectionMode.Gateway)
             {
                 this.StoreModel = this.GatewayStoreModel;
+            }
+            // Change it to this.ConnectionPolicy.ConnectionMode == ConnectionMode.LiteClient when LiteClient is supported.
+            else if (this.isLiteClientEnabled)
+            {
+                this.StoreModel = new ThinClientStoreModel(
+                    endpointManager: this.GlobalEndpointManager,
+                    this.sessionContainer,
+                    (Cosmos.ConsistencyLevel)this.accountServiceConfiguration.DefaultConsistencyLevel,
+                    this.eventSource,
+                    this.serializerSettings,
+                    this.liteModeHttpClient,
+                    new Uri(this.liteClientEndpoint),
+                    this.ServiceEndpoint.ToString());
             }
             else
             {