--- a/Microsoft.Azure.Cosmos/src/HttpClient/CosmosHttpClientCore.cs
+++ b/Microsoft.Azure.Cosmos/src/HttpClient/CosmosHttpClientCore.cs
@@ -441,7 +441,7 @@ namespace Microsoft.Azure.Cosmos
             return !timeoutEnumerator.MoveNext(); // No more retries are configured
         }
 
-        private async Task<HttpResponseMessage> ExecuteHttpHelperAsync(
+        public async Task<HttpResponseMessage> ExecuteHttpHelperAsync(
             HttpRequestMessage requestMessage,
             ResourceType resourceType,
             CancellationToken cancellationToken)